pyproject.toml
[project]
name = "wishiwashi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.45.0",
    "py-cord>=2.6.0",
    "python-dotenv>=1.0.1",
]
